@@ -32,21 +32,22 @@
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 
         <!-- Dependency versions -->
-        <jjwt.version>0.9.1</jjwt.version>
-        <jaxb-api.version>2.3.1</jaxb-api.version>
-        <validation-api.version>2.0.1.Final</validation-api.version>
+        <io.jsonwebtoken.version>0.11.5</io.jsonwebtoken.version>
+        <javax.xml.bind.version>2.3.1</javax.xml.bind.version>
+        <jakarta.validation.version>3.0.2</jakarta.validation.version>
+        <jakarta.annotation.version>2.1.1</jakarta.annotation.version>
+
         <mapstruct.version>1.5.3.Final</mapstruct.version>
         <lombok-mapstruct-binding.version>0.2.0</lombok-mapstruct-binding.version>
-        <maven-compiler-plugin.version>3.8.1</maven-compiler-plugin.version>
-        <lombok.version>1.18.22</lombok.version>
-        <spring-boot.version>2.7.5</spring-boot.version>
-        <spring-security.version>5.7.3</spring-security.version>
+        <maven-compiler-plugin.version>3.10.1</maven-compiler-plugin.version>
 
+        <lombok.version>1.18.24</lombok.version>
+        <spring-boot.version>3.0.2</spring-boot.version>
+        <spring-test.version>6.0.1</spring-test.version>
 
-        <io.springfox.version>3.0.0</io.springfox.version>
-        <org.springdoc.version>1.6.14</org.springdoc.version>
+        <io.swagger.parser.v3.version>2.1.11</io.swagger.parser.v3.version>
+        <org.springdoc.version>2.0.2</org.springdoc.version>
         <jackson-databind-nullable.version>0.2.0</jackson-databind-nullable.version>
-        <jsr250-api.version>1.0</jsr250-api.version>
     </properties>
 
     <dependencyManagement>
