@@ -1,4 +1,4 @@
-package mk.ukim.finki.coding_helper.application;
+package mk.ukim.finki.coding_helper;
 
 
 import org.springframework.boot.SpringApplication;
@@ -6,13 +6,8 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.data.neo4j.repository.config.EnableNeo4jRepositories;
 
 
-@SpringBootApplication(scanBasePackages = {
-    "mk.ukim.finki.coding_helper.integration",
-    "mk.ukim.finki.coding_helper.core",
-    "mk.ukim.finki.coding_helper.rest",
-    "mk.ukim.finki.coding_helper.application"
-})
-@EnableNeo4jRepositories("mk.ukim.finki.coding_helper.integration.repository")
+@SpringBootApplication
+@EnableNeo4jRepositories
 public class CodingHelperApplication {
   public static void main(String[] args) {
     SpringApplication.run(CodingHelperApplication.class, args);
